fix to remove windows warning
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@114454 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 610eb61b134c2bf14433bdf4110a26b185294bbc

+++ Lbcom/Velo/VeloAlgorithms/src/VeloSimTell1ClusterMaker.h
@@ -91,8 +91,8 @@ private:
   std::string m_rawEventLoc;
   int m_hitThresholdValue;
   int m_lowThresholdValue;
-  float m_hitSigmas;
-  float m_incSigmas;
+  double m_hitSigmas;
+  double m_incSigmas;
   VeloTELL1::u_int32_t m_sumThresholdValue;
   VeloTELL1::ThresholdsVec m_hitThresholds;
   VeloTELL1::ThresholdsVec m_lowThresholds;
